Feat user leader (#22)
* Logic for leader election

* think its working

* replication working

* consistancy seems to work

* feat: initial gw leader support

* feat: timestamps between gateways

* feat: implement gateway synchronization

Requests can be synchronized between gateways using the "synchronized":
true config on a per endpoint basis

* think its working

* feat: pass primairy host in proxy

* fix: send primairy back to server

* feat: redis replication

* feat: add redis publish support

* fix: redis replication

* feat: updates applying better

* feat: added change to update account

* fix: party code to 6 digits

---------

+++ userService/src/controllers/signup.ts
@@ -4,40 +4,61 @@ import { User, UserSignup } from "../models/User";
 import { validationResult } from "express-validator";
 import { createUserWithEmailAndPassword } from "firebase/auth";
 import { firebaseAuth } from "../firebase/firebaseConfig";
+import {updateReplicas} from "../sync/updateReplicas";
+import store from "../sync/store";
 
 export const signup = (req: Request, res: Response) => {
-  const valResult = validationResult(req);
-  if (!valResult.isEmpty()) {
-    return res.status(300).send(valResult.array());
+    const syncStore = store.getInstance()
+    console.log(`Signup request, isleader=${syncStore.isLeader()}`)
+  if (syncStore.getHostname() === req.get('X-Gateway-Leader')) {
+      const valResult = validationResult(req);
+      if (!valResult.isEmpty()) {
+          return res.status(300).send(valResult.array());
+      }
+
+      const user: UserSignup = req.body;
+
+      createUserWithEmailAndPassword(firebaseAuth, user.email, user.password)
+          .then((r) => {
+              r.user
+                  .getIdToken()
+                  .then((token) => {
+                      res.header("x-auth-token", token);
+                  })
+                  .catch((e) => {
+                      res.status(500).send(e.message);
+                  });
+
+              const newUser: User = {
+                  _id: r.user.uid,
+                  displayName: user.displayName,
+                  address: user.address,
+              };
+              createUser(newUser)
+                  .then((r) => {
+                      res.status(200).send("User Created");
+                      if (syncStore.isLeader()) updateReplicas(req, newUser);
+                  })
+                  .catch((e) => {
+                      console.log(e)
+                      r.user.delete().finally(() => res.status(500).send(e.message));
+                  });
+          })
+          .catch((e) => {
+              res.status(500).send(e.message);
+          });
+  } else {
+      const user: User = req.body;
+
+      if (syncStore.isDBReady()) {
+          createUser(user)
+              .then((r) => {
+                  res.status(200).send("User Created");
+              }).catch((err) => {
+                  console.log(err)
+              })
+      } else {
+          syncStore.pushToSignupQueue(user);
+      }
   }
-
-  const user: UserSignup = req.body;
-
-  createUserWithEmailAndPassword(firebaseAuth, user.email, user.password)
-    .then((r) => {
-      r.user
-        .getIdToken()
-        .then((token) => {
-          res.header("x-auth-token", token);
-        })
-        .catch((e) => {
-          res.status(500).send(e.message);
-        });
-
-      const newUser: User = {
-        _id: r.user.uid,
-        displayName: user.displayName,
-        address: user.address,
-      };
-      createUser(newUser)
-        .then((r) => {
-          res.status(200).send("User Created");
-        })
-        .catch((e) => {
-            r.user.delete().finally(() => res.status(500).send(e.message));
-        });
-    })
-    .catch((e) => {
-      res.status(500).send(e.message);
-    });
 };
